requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tzdata>=2024.1
certifi>=2024.1
